Add files via upload
In this website I have created a To do List that allows users to add the tasks they want to and can also edit on them. The main features include adding task button, remove button , task undo button and task reset button to reset the task list further, also the when clicked on tasks in the list, it is scratched out to show that you have completed that particular task. For this website I have used HTML, CSS and Vanilla javascript majorly and also made it look minimalistic and visually pleasing!!